Extract subclass from not ideal scope
